Add basic meson support
Dependencies are handled by meson's wrap manager, checking system deps
first, and falling back to building in tree.

mio was only used to mmap a single source file, and is was not easily
available, so I eliminated it as a dependency; files are no longer
mapped, but read in completely.

Signed-off-by: Duncan McBain <git@mail.duncanmcba.in>