Change environment implementation
The environment was previously implemented with a reference to its
parent. This was slightly scary as it required having a pointer to the
previous environment which was effectively optional, which was
implemented previously through an "Observer" class (i.e., a raw
pointer). The handling around changing the state of what the current
environment of the Interpreter was when looking for a variable was
somewhat complex and had implementation bugs.

The new method is instead to have a vector of environments, starting
with the program scope in the Interpreter constructor. Any time there is
a new scope in the program, a new Environment is popped onto the back of
the vector, then used for the duration of that scope. When looking for
variables, it scans through the Environments in reverse.

This appears to work much more nicely and has fixed all the variable
handling bugs I have tested so far. It's cleaner.