Fix pernicious bug in Scanner.cpp
cxx_loxi was convinced that newlines secretly in fact were a way of
saying that we were starting to parse a string. Now we break properly to
avoid the fall-through behaviour that is leaned on elsewhere in the
switch statement but is unhelpful here. This doesn't show up when using
repl, as it can't yet handle the user typing blocks across multiple
lines.