Add function and call parsing and interpretation
This patch is large and touches many files to fully incorporate the
changes required to be able to parse and interpret functions in Lox.

cxx-lox has a new set of classes - Callables, Functions and some
Builtins. The Function and the Builtins inherit from Callable. The
two builtins are `now()` and `print()`, which record the current
time and print ExprResults, respectively.

The new Expr type is the "Call", which is anything followed by
a set of parentheses. Obviously only real functions can be called,
and an error is reported when something that is not a function has
this syntax.