Optimise environment element lookup
The Environments were looking for the element in the map twice, so now
search once then re-use the results. This improves performance on a
program that finds the Fibonacci numbers less than 15 million by 100us,
or about 15%, consistently.

Many thanks for @Fmarno for this change who spotted and suggested a fix
for this mistake.