Add support for printing Callables
These things (normally functions for now) have names and can be printed
for debugging or similar. Slightly restructured code to fix recursive
include problems.

While doing this, @rafbiels found that the static assert was incorrect
in that it was always *true*, not false. Switching to && instead of ||
then caused other problems that I can't reproduce outwith this code.
I guess that it is somehow forced to instantiate by the overload set
mechanism. Therefore, this has been removed and I am accepting that
if I forget to add print support for some new type, I will get
nasty build errors, and that's OK.

Fixed the meson build (kind of) while I'm at it.