Initialise timer on builtin creation
Previously, the timer would start the first time you called now().
This is fine, but is a bit strange. It was also static across the
entirety of cxx_loxi which is likewise strange. Now, it is set
per-interpreter instance, and will also start counting when the
interpreter is created, which probably makes more sense.